requirements.txt
paramiko
google-cloud-storage
google-cloud-secret-manager
flask
gunicorn
